--- django/core/templates/widgets/s3fileinput.html
@@ -1,1 +1,0 @@
-{{widget}}
