+++ django/joist/models.py
@@ -1,14 +1,13 @@
 from django.contrib.auth import get_user_model
 from django.db import models
 from django.utils import timezone
-
-from core.models import CollisionSafeFileField
+from s3widget.models import S3FileField
 
 
 class Blob(models.Model):
     created = models.DateTimeField(default=timezone.now)
     creator = models.ForeignKey(get_user_model(), on_delete=models.CASCADE)
-    resource = CollisionSafeFileField()
+    resource = S3FileField()
 
     def __str__(self):
         return str(self.resource)
