+++ django/joist/views.py
@@ -1,16 +1,15 @@
 from django.contrib.auth.models import User
 from django.http.response import HttpResponseBase
+from django.shortcuts import HttpResponseRedirect, render
+from django.views import generic
 from rest_framework import viewsets
 from rest_framework.decorators import api_view, parser_classes
 from rest_framework.parsers import JSONParser
 from rest_framework.request import Request
 from rest_framework.response import Response
 
-from django.shortcuts import render, HttpResponseRedirect
-from django.views import generic
-
-from .models import Blob
 from .forms import BlobForm
+from .models import Blob
 from .serializers import BlobSerializer
 
 
@@ -46,7 +45,7 @@ class DetailView(generic.DetailView):
     template_name = 'blob/detail.html'
 
 
-def newBlob(request: Request) -> HttpResponseBase:
+def new_blob(request: Request) -> HttpResponseBase:
     form = BlobForm(request.POST, request.FILES) if request.method == 'POST' else BlobForm()
     if request.method == 'POST':
         # check whether it's valid:
