+++ django/s3widget/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class S3WidgetConfig(AppConfig):
+    name = 's3widget'
