+++ django/s3widget/models.py
@@ -1,9 +1,11 @@
 from uuid import uuid4
 
-from django.db import models
+from django.db.models import FileField
+
+from .widgets import S3FormFileField
 
 
-class CollisionSafeFileField(models.FileField):
+class S3FileField(FileField):
     description = 'A file field which is uploaded to <randomuuid>/filename.'
 
     def __init__(self, *args, **kwargs):
@@ -14,3 +16,8 @@ class CollisionSafeFileField(models.FileField):
     @staticmethod
     def uuid_prefix_filename(instance, filename):
         return f'{uuid4()}/{filename}'
+
+    def formfield(self, **kwargs):
+        return super().formfield(
+            **{'form_class': S3FormFileField, 'max_length': self.max_length, **kwargs}
+        )
