+++ django/s3widget/templates/s3fileinput/s3fileinput.html
@@ -1,0 +1,1 @@
+{{widget}}
