+++ django/tox.ini
@@ -3,6 +3,7 @@ envlist =
     lint,
     typing,
 
+
 [testenv:lint]
 basepython = python3
 skipsdist = true
@@ -24,6 +25,7 @@ commands =
 basepython = python3
 skipsdist = true
 skip_install = true
+whitelist_externals = bash
 deps =
     mypy
     django-stubs
@@ -32,7 +34,7 @@ deps =
 setenv =
     PYTHONPATH = {env:PYTHONPATH:}{:}{envsitepackagesdir}{:}{toxinidir}
 commands =
-    mypy {posargs:.}
+    bash -c "source ./.env && mypy {posargs:.}"
 
 [testenv:format]
 basepython = python3
